==> 02.2-address-of-account.ipynb <==
from sqlalchemy import create_engine
import pandas as pd

from config import db_connection_string

engine = create_engine(db_connection_string)

## match contribution and tx

df = pd.read_sql_query("""
SELECT 
	`t1`.`id` AS `contribution`,
	`t1`.`account`,
    `t1`.`grant`,
	`t3`.`from`,
	`t3`.`tx`
FROM
    `contribution` AS `t1`
INNER JOIN
    `grant` AS `t2` 
	ON 
		`t2`.`id` = `t1`.`grant`
INNER JOIN
    `tx` AS `t3` 
    ON 
		`t3`.`to` = `t2`.`address`
        AND `t3`.`symbol` = `t1`.`symbol`
        AND `t3`.`amount` = `t1`.`amount_to_grant`
        AND ABS(TIMESTAMPDIFF(MINUTE, `t3`.`created_on`, `t1`.`created_on`)) <= 5
""", engine)

df.shape, df['tx'].nunique()

df['n_account_in_tx'] = df.groupby('tx')['account'].transform('nunique')
df['contribution_exists_n_times'] = df.groupby('contribution')['contribution'].transform(len)
df['tx_exists_n_times_in_grant'] = df.groupby(['grant', 'tx'])['tx'].transform(len)


## tx (only matching one account)

account_address_df = df[df['n_account_in_tx']==1].groupby(['account', 'from']).size().reset_index(name='tx_count')

engine.execute("""
    CREATE TABLE `account_address`(
        `account` VARCHAR(300) NOT NULL,
        `address` VARCHAR(256) NOT NULL,
        `tx_count` INT NOT NULL,
        KEY `account_index`(`account`),
        KEY `address_index`(`address`)
    )
""")

account_address_df.rename(columns={'from':'address'}).to_sql("account_address", engine, index=False, if_exists="append")

## other interesting things

### addresses used by multiple accounts

account_address_df['from'].value_counts()

(account_address_df['from'].value_counts()>1).sum()

### accounts with multiple addresses

account_address_df['account'].value_counts()

(account_address_df['account'].value_counts()>1).sum()

### duplicate contributions?  
These contributions can only match one tx, and most of them been created at alomost the same time, maybe they are duplicate records?

df[
    (df['tx_exists_n_times_in_grant'] > 1) & 
    (df['n_account_in_tx'] == 1) & 
    ~df['tx'].isin(df[df['contribution_exists_n_times'] > 1]['tx'].unique())
]
