==> 03.2-contribution-IoU.ipynb <==
import json
import pandas as pd
from tqdm.notebook import tqdm
from functools import partial
from sqlalchemy import MetaData, Table, create_engine
from sqlalchemy.sql import text

from config import db_connection_string
from utils import tx_count_between_accounts, tx_count_between_grants
from iou_calculator import IoUCalculator


engine = create_engine(db_connection_string)

with open('raw_data/contributions.json', 'r') as file:
    contributions = json.load(file)

contributions = [x for x in contributions if x['grant']!='86' and x['success']]

def generate_suspicious_ratio_df(pairs, objects, iou_threshold_list, min_contribution_count_list):
    return pd.DataFrame(
        [
            [
                iou_threshold, min_contribution_count, 
                (
                    str(sum(1 if len(x['contributions']) >= min_contribution_count else 0 for x in objects.values())) + ',' +
                    str(len({pair[i] for pair in pairs for i in range(2) if pair[2] >= iou_threshold and pair[3] >= min_contribution_count})) + ',' +
                    str(len({pair[i] for pair in pairs for i in range(2) if pair[2] >= iou_threshold and pair[3] >= min_contribution_count and pair[4] >= 1}))
                )
            ]
            for iou_threshold in iou_threshold_list
            for min_contribution_count in min_contribution_count_list
        ],
        columns=['iou_threshold', 'min_contribution_count', 'suspicious_account_ratio']
    ).pivot(index='iou_threshold', columns='min_contribution_count', values='suspicious_account_ratio')


## account suspicious ratio(grant)

calculator = IoUCalculator(
    contributions, lambda x: x['subscription']['contributor_profile'], lambda x: x['grant'],
    lambda x, y: tx_count_between_accounts(x, y, engine)
)
pairs = calculator.compute_suspicious_pairs(0.5, 1)
generate_suspicious_ratio_df(pairs, calculator.objects, [round(x * 0.1, 1) for x in range(5, 11)], [1, 3, 5, 10, 15, 20])

## account suspicious ratio(grant+date)

calculator = IoUCalculator(
    contributions, lambda x:x['subscription']['contributor_profile'], lambda x:(x['grant'], x['created_on'][:10]),
    lambda x, y: tx_count_between_accounts(x, y, engine)
)
pairs = calculator.compute_suspicious_pairs(0.5, 1)
generate_suspicious_ratio_df(pairs, calculator.objects, [round(x * 0.1, 1) for x in range(5, 11)], [1, 3, 5, 10, 15, 20])


## grant suspicious ratio(account)

calculator = IoUCalculator(
    contributions, lambda x:x['grant'], lambda x:x['subscription']['contributor_profile'],
    lambda x, y: tx_count_between_grants(x, y, engine)
)
pairs = calculator.compute_suspicious_pairs(0.5, 1)
generate_suspicious_ratio_df(pairs, calculator.objects, [round(x * 0.1, 1) for x in range(5, 11)], [1, 3, 5, 10, 15, 20])


## grant suspicious ratio(account+date)

calculator = IoUCalculator(
    contributions, lambda x:x['grant'], lambda x:(x['subscription']['contributor_profile'], x['created_on'][:10]),
    lambda x, y: tx_count_between_grants(x, y, engine)
)
pairs = calculator.compute_suspicious_pairs(0.5, 1)
generate_suspicious_ratio_df(pairs, calculator.objects, [round(x * 0.1, 1) for x in range(5, 11)], [1, 3, 5, 10, 15, 20])
