==> 03.3-grant-with-contributors-only-contribute-to-afew-grants.ipynb <==
import json
import pandas as pd
from tqdm.notebook import tqdm
from functools import partial
from sqlalchemy import MetaData, Table, create_engine
from sqlalchemy.sql import text

from config import db_connection_string
from utils import tx_count_between_accounts, tx_count_between_grants
from iou_calculator import IoUCalculator


engine = create_engine(db_connection_string)

engine.execute("""
    CREATE TEMPORARY TABLE `account`
    (INDEX `account_index` (`account`))
    SELECT 
        `account`, 
        COUNT(DISTINCT `grant`) AS `distinct_grant_count`
    FROM `contribution`
    WHERE `grant`!='86'
    GROUP BY `account`;
            
    CREATE TEMPORARY TABLE `grant_contributor`
    (INDEX `account_index` (`account`))
    SELECT DISTINCT 
        `grant`, 
        `account`
    FROM `contribution`
    WHERE `grant`!='86';
""")


grant_contributor_df = pd.read_sql_query("""
    SELECT 
        `grant_contributor`.`grant`, 
        `account`.`account` AS `contributor`, 
        `account`.`distinct_grant_count` AS `contributor_distinct_grant_count`
    FROM `grant_contributor`
    INNER JOIN `account` ON `account`.`account`=`grant_contributor`.`account`
""", engine)


pd.DataFrame([
    [
        contributed_grants_less_than_equal,
        round(ratio / 10, 1),
        (
            grant_contributor_df.groupby('grant').agg(
                ratio=('contributor_distinct_grant_count', lambda x: (
                    x <= contributed_grants_less_than_equal).sum() / x.shape[0])
            )['ratio'] >= ratio / 10
        ).sum()
    ]
    for contributed_grants_less_than_equal in range(1, 11)
    for ratio in range(1, 11)
], columns=['account contributed grants<=', 'matched account ratio>=', 'count']).pivot(index='account contributed grants<=', columns='matched account ratio>=', values='count')
